Napoleonic France Flavor Additions
Hi. Took a bit for some sort of an Update, but alas it's here even if it's not proper.

In any case, I'm working towards a 1.0 where every GP has flavor. It might take a bit (especially since I'm focusing on other projects) but otherwise it should come out either soon or later. Stuff like this takes time.

In any case, this Mod is still very untested so expect issues. Otherwise, enjoy.