Russian Flavor Update
Ok so Russia. I'm gonna get to the point fast.

This might be the most I've given a singular Country to be honest. I wasn't intending on giving this much, but hey who am I to complain?

Anyways, we're almost done. The Ottomans and Brazil remain.

I promise Version 1.0 sometime soon-ish. IF AND ONLY IF I ACTUALLY WORK CONSISTENTLY ON IT

anyways yeah that's about it