holy crap i'm done
Yeah that's right, I've actually done it.

So I actually played the Mod. Turns out that the only real thing I needed to change heavily was any of the Events relating to the Liberal Revolutions file. No joke that made the mod very much unfun.

Besides that, most things worked as they should, though I did go ahead and tweak some things in order to make sure it's a semi-balanced Playthrough.

I was suprised that Dismantlements worked. I was sure that those were going to not function.

What else... oh right, Flavor. Brazil and the Ottomans have their flavor now, meaning that every Great Power has things to do (and hell some Secondaries do as well).

Now with all of this done and over with, I'll go ahead and put out the 1.0 release. It took a while, but we're at the end now.